# feat(treesitter): add support for wasm parsers