### Problem

Treesitter parsers are currently distributed as compiled shared objects which need to be compiled by the users (with the help of nvim-treesitter).

### Solution

Allow loading wasm built parsers.

Notes:
- For now this is an optional build-time feature (via `ENABLE_WASMTIME`) which will be dynamically linked.
  - Builds with `make CMAKE_EXTRA_FLAGS=-DENABLE_WASMTIME=ON DEPS_CMAKE_FLAGS=-DENABLE_WASMTIME=ON`
- Treesitter requires being built with `TREE_SITTER_FEATURE_WASM` and also requires `wasmtime` at build time. So no `[uv_]dlopen` 😢 
- `wasmtime` has includes that require C11 (`static_assert`).
- To reduce re-compilation times, install `sccache` and build with `RUSTC_WRAPPER=sccache make ...`.

### Status

Waiting for the next tree-sitter release with the commit we're pulling in here to allow us to build against a wasmtime release with the build changes required to avoid unacceptable regressions on macOS. (Ideally, that release would also be built against the same version we use, but that is blocked by the same build system changes; luckily so far the versions have not become incompatible for our purpose.)