@clason 2024-04-23
@lewis6991 wasmtime 20.0.0 just got released (and is already updated on homebrew). Could you test your PR again to see if this is breaking?