@dundargoc 2024-04-23
We are not using the release tags, where are using random commit on main. We can bump the commit version of course.