@clason 2024-04-23
Ah, right. Yes, that would be interesting (to see _how_ breaking the wasmtime update velocity is for us). 

But can't we use release tags (not releases)? I think that would be more meaningful and easier to handle than random commits.