@dundargoc 2024-04-23
> But can't we use release tags (not releases)? 

Nah, I tried. `wasm.h` header isn't included in the release tags for some reason, which is needed. `wasm.h` is from https://github.com/WebAssembly/wasm-c-api/blob/main/include/wasm.h that they have vendored on the main branch.

I can ask in wasmtime repo how we're meant to use a release tag since `wasm.h` isn't included in their release tags.