@clason 2024-04-23
I just checked; https://github.com/bytecodealliance/wasmtime/archive/v20.0.0.tar.gz _does_ contain `wasm.h`?