@dundargoc 2024-04-23
> I just checked; [bytecodealliance/wasmtime@`v20.0.0`.tar.gz](https://github.com/bytecodealliance/wasmtime/archive/v20.0.0.tar.gz) _does_ contain [`wasm.h`](https://github.com/bytecodealliance/wasmtime/blob/v20.0.0/crates/c-api/include/wasm.h?rgh-link-date=2024-04-23T10%3A56%3A20Z)? (release tag != release)

The previous releases didn't, so thanks for the gaslight wasmtime.