@clason 2024-04-23
That's right; they switched from a submodule to vendoring it for v20.0.0. Sorry, I should have checked older versions as well.