@dundargoc 2024-04-23
Maybe we should add a make variable that passes cmake args to both deps and neovim, `make CMAKE_EXTRA_FLAGS=-DENABLE_WASMTIME=ON DEPS_CMAKE_FLAGS=-DENABLE_WASMTIME=ON` is a bit much lol