@dundargoc 2024-04-23
TODO: make it work on windows