@dundargoc 2024-04-26
Soft-dependent on https://github.com/bytecodealliance/wasmtime/pull/8490, aka would simplify some things but not a blocker.