@justinmk 2024-04-26
> thoughts on merging now or just wait until the release?

No objection, because this doesn't introduce surface area that would be hard to change. If it causes any problems with the 0.10 release we can always hotfix stuff. That's different than e.g. a new API that would then need a deprecation cycle.