@dundargoc 2024-05-04
This works on windows release-builds. The debug build fails with linker warnings. It would be good to get that to work, but not a blocker IMO. We can ship it as is for the time being when we can.