@clason 2024-05-04
Preliminary testing indicates that building nvim (and tree-sitter) with wasmtime has a significant performance impact, even without using wasm parsers: startuptime increases between 50% and 200% (depending on config and how you measure); the cost is somewhere in the Lua API stack.

(Comparing `--clean` shows little to no difference, so this is not related to the binary size increasing by ~20Mb due to the static linkage of wasmtime.)