@lewis6991 2024-05-04
If `--clean` shows no difference then I find that a bit sus as no extra code is called in `treesitter.c`.