@clason 2024-05-04
It _is_ sus, which is why I report it here as preliminary information. Obviously we need to profile where the increase comes from; it's entirely possible that it's from tree-sitter itself if built with wasmtime enabled.