@lewis6991 2024-05-04
`tslua_free` should only be needed for Asan.