@clason 2024-05-04
Yeah, it's not that (unsurprisingly); it was just an example of extra code in this PR. I have a hard time tracing the `--startuptime` differences to a specific point in my config.