@lewis6991 2024-05-04
You could try reverting all the C changes to see if the issue persists? That would rule out whether it's our code or not.