@clason 2024-05-04
That makes no difference. But it's weird that the startup time for (nearly) _all_ plugins increased across the board, including for `gitsigns` which has no treesitter interaction. Maybe the issue is a build flag for one of the other dependencies, or an optimization that fails with wasmtime statically linked? (Grasping at straws here, obviously.)