@clason 2024-05-04
After more voodoo debugging, here's a few findings:
1. performance hit seems to scale with binary size
2. using (full) `lto = true` drops the size by ~10 Mb and reduces the impact to ~10ms (from ~35ms)
3. nothing else (`strip  = true`, `opt-level = 's'`) seems to make a difference.

This is on macOS, by the way; also PUC Lua (or `jit.off`) may behave differently.