@lewis6991 2024-05-04
LTO generally makes a big difference on macos anyway so that makes sense. Maybe the change in binary size is causing the CPU caches to not work as they did before, and LTO just mitigates that.

This is where statically linking everything is beginning to show its costs. Could we maybe somehow link wasmtime dynamically (as a shared lib), but by using a shared lib we build? A sort of half way house between static and dynamic linking? Or this that a fake thing I just made up? @dundargoc