@clason 2024-05-04
That was my idea as well (it's what `/usr/local/lib/nvim/` is for, anyway), but @dundargoc mentioned that this is (borderline) impossible?