@lewis6991 2024-05-04
Maybe we just accept the hit and rely on LTO (and potentially any other magic flags).

Having wasm available opens some pretty big doors in general if we still want to add support for wasm plugins.