@clason 2024-05-04
> Maybe we just accept the hit and rely on LTO (and potentially any other magic flags).

Even with LTO (which increases build time by a factor of two, and requires hackery or forkery), I don't think it's acceptable to just shrug our shoulders and enable it by default. I note again that this is all without touching a single tree-sitter parser, let alone spooling up a wasmtime engine! An optional feature that negatively affects _all_ operations across the board is not a good trade-off. (And I'm saying that as the person who is most aggressively pushing for wasm parsers...)