@dundargoc 2024-05-04
> Could we maybe somehow link wasmtime dynamically (as a shared lib)

Yes, that should be possible.

> but by using a shared lib we build and ship with nvim (nvim-wasmtime)? A sort of half way house between static and dynamic linking?

We can build wasmtime as a shared library. The problem comes from that we can't use shared libraries for our releases, meaning that we won't be able to use wasmtime on our releases. Unsure if that's an acceptable trade, but I don't mind testing it out.