@dundargoc 2024-05-04
I've added some different flags and I don't notice a difference now. Please test again.