@lewis6991 2024-05-04
> Even with LTO (which increases build time by a factor of two, and requires hackery or forkery to even enable), I don't think it's acceptable to just shrug our shoulders and enable it by default.

The build time and binary size was mostly a given, no? I began this PR with that understanding.

The load time is the main concern for me, but ~10ms seems shruggable given the context. Especially if that means long term we can avoid users loading binary files, which I see as a huge security risk, amongst all the other issues.

The hackery and forkery should be temporary?

Just to note, I'm not aiming to merge this for 0.10, in case that wasn't understood. My mindset for this is for the longer term.