@clason 2024-05-04
My mindset as well. But it's _not_ just about load time; it looks to impact LuaJIT across the board (since the time difference scales strongly with complexity of the config). It's ok if the cost is associated with use, but that is very much not the case here.

> The build time and binary size was mostly a given, no? I began this PR with that understanding.

Yes, but we can (and should) still try to minimize that, given that they're quite egregious (increase by a factor of ~4).

> The hackery and forkery should be temporary?

That depends on whether wasmtime is willing to enable (much) more aggressive optimizations by default. I'm a bit skeptical about this (but surprises happen).