@clason 2024-06-06
Update: That works (including actually loading a wasm parser), and cuts down wasmtime build by about half (60s -> 160s full nvim release build on my potato Intel mac). Binary size goes from 6.7 Mbyte to 12 Mbyte, startup from 30ms -> 30.5ms.

We could now even drop the full LTO from the build (60s -> 107s, 6.7 Mbyte -> 15 Mbyte, 30ms -> 31.5ms startup), but why say no to free performance? ;)